+++ CHANGELOG.md
@@ -21,6 +21,7 @@ for further information about branching and tagging conventions.
 - A bug with possibility select date period start after date period end
   in StatisticsFilters
 - A bug with wrong displaying long email in Passport
+- A possibility select expiration date of asset before current time
 
 ## [1.7.0-rc.3] - 2020-03-25
 #### Fixed